requirements.txt
django
django-heroku
djangorestframework
gunicorn
psycopg2-binary
django-cors-headers
